feat: getFestivityDetail api 구현

+++ src/main/java/com/example/busanzipback/domain/map/repository/FestivityRepository.java
@@ -20,5 +20,5 @@ public interface FestivityRepository extends JpaRepository<Festivity, Long> {
 		nativeQuery = true)
 	List<Festivity> findNearbyLocations(@Param("longitude") Double longitude, @Param("latitude") double latitude);
 
-	Optional<Festivity> findById(Long festivityId);
+	Optional<Festivity> findById(Integer festivityId);
 }
